done, report made fibs calculated

+++ CS162/L10/FibonacciNR.cpp
@@ -6,7 +6,7 @@ FibonacciNR::FibonacciNR() {
 
 }
 
-FibonacciNR::FibonacciNR(const int &n):n_(&n){
+FibonacciNR::FibonacciNR(const unsigned long long &n):n_(&n){
 
 }
 
@@ -14,13 +14,13 @@ FibonacciNR::~FibonacciNR() {
 
 }
 
-int FibonacciNR::Fibonacci(const int &n) {
-   int first =0;
-   int second=1;
-   int counter=2;
+unsigned long long FibonacciNR::Fibonacci(const unsigned long long &n) {
+   unsigned long long first =0;
+   unsigned long long second=1;
+   unsigned long long counter=2;
    while(counter < n)
    {
-       int temp=second;
+       unsigned long long temp=second;
        second=first+second;
        first=temp;
        ++counter;
@@ -32,6 +32,6 @@ int FibonacciNR::Fibonacci(const int &n) {
 }
 
 void FibonacciNR::PrintFibonacci() {
-   const int result = Fibonacci(*n_);
+   const unsigned long long result = Fibonacci(*n_);
   std::cout << *n_ << "th fibonaci number: " << result << std::endl;
 }
